- GUI RPC: switching to the new XML parser     (which won't parse a double as an int)     revealed a type mismatch in FILE_TRANSFER::next_request_time     between client and server.
svn path=/trunk/boinc/; revision=25125

+++ lib/gui_rpc_client.h
@@ -308,8 +308,8 @@ public:
     bool pers_xfer_active;
     bool xfer_active;
     int num_retries;
-    int first_request_time;
-    int next_request_time;
+    double first_request_time;
+    double next_request_time;
     int status;
     double time_so_far;
     double bytes_xferred;
